Extract FileNamingConvention mixin, fix dot escaping, use kopf.dev prefix
- Extract _escape() and _build_filename() into a FileNamingConvention
  mixin in conventions.py, shared by FileProgressStorage and
  FileDiffBaseStorage.
- Fix dot escaping: only protect against '..' as a standalone path
  component (path traversal), keep normal dots in place for names
  like 'my.app.v1'.
- Change default prefix from 'kopf.zalando.org' to 'kopf.dev' for
  new file-based storage classes.
