+++ kopf/_cogs/configs/progress.py
@@ -42,7 +42,6 @@ import abc
 import copy
 import json
 import pathlib
-import urllib.parse
 from collections.abc import Collection, Mapping
 from typing import Any, TypedDict, cast
 
@@ -374,7 +373,7 @@ class StatusProgressStorage(ProgressStorage):
         return essence
 
 
-class FileProgressStorage(ProgressStorage):
+class FileProgressStorage(conventions.FileNamingConvention, ProgressStorage):
     """
     State storage in YAML files on a shared filesystem or pod volume.
 
@@ -405,31 +404,12 @@ class FileProgressStorage(ProgressStorage):
             self,
             *,
             path: str | pathlib.Path,
-            prefix: str = 'kopf.zalando.org',
+            prefix: str = 'kopf.dev',
             touch_key: str = 'touch-dummy',
     ) -> None:
-        super().__init__()
-        self._path = pathlib.Path(path)
+        super().__init__(path=path, file_suffix='progress')
         self.prefix = prefix
         self.touch_key = touch_key
-
-    @staticmethod
-    def _escape(value: str) -> str:
-        """Percent-encode a value for safe use in filenames."""
-        # urllib.parse.quote never encodes dots (unreserved in RFC 3986),
-        # so we replace them manually to prevent '..' path traversal.
-        return urllib.parse.quote(value, safe='-_~').replace('.', '%2E')
-
-    def _build_filename(self, body: bodies.Body) -> pathlib.Path | None:
-        namespace = body.get('metadata', {}).get('namespace')
-        name = body.get('metadata', {}).get('name')
-        uid = body.get('metadata', {}).get('uid')
-        if not name or not uid:
-            return None
-        safe_name = self._escape(name)
-        safe_uid = self._escape(uid)
-        prefix = f'{self._escape(namespace)}-' if namespace else ''
-        return self._path / f'{prefix}{safe_name}-{safe_uid}.progress.yaml'
 
     def fetch(
             self,
